few fixes from merges
git-svn-id: https://svn.atnf.csiro.au/difx@313 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/branches/Perth_merge/src/configuration.cpp
@@ -618,7 +618,7 @@ void Configuration::processConfig(ifstream * input)
     getinputline(input, &line, "NUM CHANNELS");
     configs[i].numchannels = atoi(line.c_str());
     getinputline(input, &line, "CHANNELS TO AVERAGE");
-    configs[i].channelstoaverage = atof(line.c_str());
+    configs[i].channelstoaverage = atoi(line.c_str());
     getinputline(input, &line, "OVERSAMPLE FACTOR");
     configs[i].oversamplefactor = atoi(line.c_str());
     configs[i].independentchannelindex = i;
